requirements.txt
asgiref==3.5.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-rest-auth==2.2.4
Django==4.0
django-cors-headers==3.11.0
django-countries==7.3.2
django-filter==21.1
django-jazzmin==2.5.0
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
django-split-settings==1.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
django-extensions==3.1.5 
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
ipython==8.2.0 
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
phonenumbers==8.12.45
platformdirs==2.5.1
psycopg2==2.9.3
PyJWT==2.3.0
pyparsing==3.0.8
python-decouple==3.6
pytz==2022.1
requests==2.27.1
requests-mock==1.9.3
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
stripe==2.71.0
tomli==2.0.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
